+++ BitbankDotNet/Entities/Order.cs
@@ -35,30 +35,30 @@ namespace BitbankDotNet.Entities
         /// 注文時の数量
         /// </summary>
         [DataMember(Name = "start_amount")]
-        public double StartAmount { get; set; }
+        public decimal StartAmount { get; set; }
 
         /// <summary>
         /// 未約定の数量
         /// </summary>
         [DataMember(Name = "remaining_amount")]
-        public double RemainingAmount { get; set; }
+        public decimal RemainingAmount { get; set; }
 
         /// <summary>
         /// 約定済み数量
         /// </summary>
         [DataMember(Name = "executed_amount")]
-        public double ExecutedAmount { get; set; }
+        public decimal ExecutedAmount { get; set; }
 
         /// <summary>
         /// 注文価格
         /// </summary>
-        public double Price { get; set; }
+        public decimal Price { get; set; }
 
         /// <summary>
         /// 平均約定価格
         /// </summary>
         [DataMember(Name = "average_price")]
-        public double AveragePrice { get; set; }
+        public decimal AveragePrice { get; set; }
 
         /// <summary>
         /// 注文日時
