+++ BitbankDotNet/Entities/Ticker.cs
@@ -12,32 +12,32 @@ namespace BitbankDotNet.Entities
         /// <summary>
         /// 現在の売り注文の最安値
         /// </summary>
-        public double Sell { get; set; }
+        public decimal Sell { get; set; }
 
         /// <summary>
         /// 現在の買い注文の最高値
         /// </summary>
-        public double Buy { get; set; }
+        public decimal Buy { get; set; }
 
         /// <summary>
         /// 過去24時間の最高値取引価格
         /// </summary>
-        public double High { get; set; }
+        public decimal High { get; set; }
 
         /// <summary>
         /// 過去24時間の最安値取引価格
         /// </summary>
-        public double Low { get; set; }
+        public decimal Low { get; set; }
 
         /// <summary>
         /// 最新取引価格
         /// </summary>
-        public double Last { get; set; }
+        public decimal Last { get; set; }
 
         /// <summary>
         /// 過去24時間の出来高
         /// </summary>
-        public double Vol { get; set; }
+        public decimal Vol { get; set; }
 
         /// <summary>
         /// 日時
